requirements.txt
fastapi
uvicorn
pymongo
pydantic
passlib
pyjwt
python-multipart
cffi
bcrypt
aiofiles